dom/media/webspeech/synth/test/file_speech_error.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1226015
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1226015</title>
  <script type="application/javascript">
    window.SimpleTest = parent.SimpleTest;
    window.info = parent.info;
    window.is = parent.is;
    window.isnot = parent.isnot;
    window.ok = parent.ok;
  </script>
  <script type="application/javascript" src="common.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1226015">Mozilla Bug 1226015</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 1226015 **/

function testFunc(done_cb) {
  var utterance = new SpeechSynthesisUtterance();
  utterance.lang = 'it-IT-error';

  speechSynthesis.speak(utterance);
  speechSynthesis.cancel();

  ok(true, "we didn't crash, that is good.")
  SimpleTest.finish();
}

// Run test with no global queue, and then run it with a global queue.
testFunc();

</script>
</pre>
</body>
</html>
